--- src/pages/HomePage/components/AlbumRatings/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AlbumRatings'
