--- src/pages/HomePage/components/AlbumsSearchField/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AlbumsSearchField'
