--- src/pages/HomePage/components/TopAlbums/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TopAlbums'
